+++ WordPress/Tests/DB/PreparedSQLUnitTest.php
@@ -7,7 +7,7 @@
  * @license https://opensource.org/licenses/MIT MIT
  */
 
-namespace WordPress\Tests\WP;
+namespace WordPress\Tests\DB;
 
 use PHP_CodeSniffer\Tests\Standards\AbstractSniffUnitTest;
 
@@ -18,8 +18,7 @@ use PHP_CodeSniffer\Tests\Standards\AbstractSniffUnitTest;
  *
  * @since   0.8.0
  * @since   0.13.0 Class name changed: this class is now namespaced.
- * @since   0.15.0 The sniff has been deprecated. This unit test file now
- *                 only tests that the deprecation warnings are correctly thrown.
+ * @since   0.15.0 This sniff has been moved from the `WP` category to the `DB` category.
  */
 class PreparedSQLUnitTest extends AbstractSniffUnitTest {
 
@@ -31,7 +30,23 @@ class PreparedSQLUnitTest extends AbstractSniffUnitTest {
 	 * @return array <int line number> => <int number of errors>
 	 */
 	public function getErrorList() {
-		return array();
+		return array(
+			3  => 1,
+			4  => 1,
+			5  => 1,
+			7  => 1,
+			8  => 1,
+			16 => 1,
+			17 => 1,
+			18 => 1,
+			20 => 1,
+			21 => 1,
+			54 => 1,
+			64 => 1,
+			71 => 1,
+			85 => 1,
+			90 => 1,
+		);
 	}
 
 	/**
@@ -42,9 +57,7 @@ class PreparedSQLUnitTest extends AbstractSniffUnitTest {
 	 * @return array <int line number> => <int number of warnings>
 	 */
 	public function getWarningList() {
-		return array(
-			1 => 1,
-		);
+		return array();
 
 	}
