+++ WordPress/ruleset.xml
@@ -61,4 +61,14 @@
 		<severity>0</severity>
 	</rule>
 
+	<!-- Prevent deprecation notice when the sniff is not explicitely included. -->
+	<rule ref="WordPress.WP.PreparedSQL.DeprecatedSniff">
+		<severity>0</severity>
+	</rule>
+
+	<!-- Prevent duplicate messages from deprecated sniff. -->
+	<rule ref="WordPress.WP.PreparedSQL.NotPrepared">
+		<severity>0</severity>
+	</rule>
+
 </ruleset>
